lib: disable REPL preview completion on nested properties
the REPL completion logic evaluates code, this is generally ok
but it can be problematic when there are objects with nested
properties since the code evaluation would trigger their potential
getters (e.g. the `obj.foo.b` line would trigger the getter of
`foo`), in order to make this not too surprising and disruptive
for the user the changes here make it so that the code completion
is disabled in previews (e.g. the REPL will only try to complete
`obj.foo.b`, thus triggering the getter of `foo`, if the user
actively asks for code completion by pressing tab)